Fix: show error msg when tx is not found
