+++ src/routes/safe/components/SafeLoadError.tsx
@@ -1,16 +1,11 @@
 import { ReactElement, useEffect } from 'react'
-import { useHistory } from 'react-router-dom'
-import { Title } from '@gnosis.pm/safe-react-components'
-import styled from 'styled-components'
-
-import Button from 'src/components/layout/Button'
 import { extractSafeAddress, WELCOME_ROUTE } from 'src/routes/routes'
 import { useDispatch } from 'react-redux'
 import removeViewedSafe from 'src/logic/currentSession/store/actions/removeViewedSafe'
+import FetchError from './FetchError'
 
 const SafeLoadError = (): ReactElement => {
   const dispatch = useDispatch()
-  const history = useHistory()
 
   useEffect(() => {
     // Remove the errorneous Safe from the list of viewed safes on unmount
@@ -19,30 +14,7 @@ const SafeLoadError = (): ReactElement => {
     }
   }, [dispatch])
 
-  const handleClick = () => {
-    history.push(WELCOME_ROUTE)
-  }
-
-  return (
-    <ErrorContainer>
-      <img src="./resources/error.png" alt="Error" />
-      <Title size="xs">This Safe couldn&apos;t be loaded</Title>
-      <Button onClick={handleClick} color="primary" size="medium" variant="contained">
-        Back to Main Page
-      </Button>
-    </ErrorContainer>
-  )
+  return <FetchError text="This Safe couldn't be loaded" buttonText="Back to Main Page" redirectRoute={WELCOME_ROUTE} />
 }
-
-export const ErrorContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  // Offset so that it is centered relative to the header
-  margin-top: -30px;
-`
 
 export default SafeLoadError
