+++ src/routes/safe/components/Transactions/TxList/TxSingularDetails.tsx
@@ -1,5 +1,5 @@
 import { ReactElement, useEffect, useState } from 'react'
-import { useParams } from 'react-router-dom'
+import { useHistory, useParams } from 'react-router-dom'
 import { Loader } from '@gnosis.pm/safe-react-components'
 import { shallowEqual, useDispatch, useSelector } from 'react-redux'
 import { TransactionDetails } from '@gnosis.pm/safe-react-gateway-sdk'
@@ -12,7 +12,6 @@ import {
   SafeRouteSlugs,
   SAFE_ROUTES,
   TRANSACTION_ID_SLUG,
-  history,
 } from 'src/routes/routes'
 import { Centered } from './styled'
 import { getTransactionWithLocationByAttribute } from 'src/logic/safe/store/selectors/gatewayTransactions'
@@ -30,12 +29,15 @@ import { Transaction } from 'src/logic/safe/store/models/types/gateway.d'
 import { currentChainId } from 'src/logic/config/store/selectors'
 import { QueueTxList } from './QueueTxList'
 import { HistoryTxList } from './HistoryTxList'
+import FetchError from '../../FetchError'
 
 const TxSingularDetails = (): ReactElement => {
   const { [TRANSACTION_ID_SLUG]: safeTxHash = '' } = useParams<SafeRouteSlugs>()
   const [fetchedTx, setFetchedTx] = useState<TransactionDetails>()
   const [liveTx, setLiveTx] = useState<{ txLocation: TxLocation; transaction: Transaction }>()
+  const [error, setError] = useState<Error>()
   const dispatch = useDispatch()
+  const history = useHistory()
   const chainId = useSelector(currentChainId)
 
   // We must use the tx from the store as the queue actions alter the tx
@@ -76,6 +78,7 @@ const TxSingularDetails = (): ReactElement => {
         txDetails = await fetchSafeTransaction(safeTxHash)
       } catch (e) {
         logError(Errors._614, e.message)
+        setError(e)
         return
       }
 
@@ -89,7 +92,7 @@ const TxSingularDetails = (): ReactElement => {
     return () => {
       isCurrent = false
     }
-  }, [safeTxHash, setFetchedTx, setLiveTx])
+  }, [history, safeTxHash, setFetchedTx, setLiveTx])
 
   // Add the tx to the store
   useEffect(() => {
@@ -111,6 +114,16 @@ const TxSingularDetails = (): ReactElement => {
     // And add it to the corresponding list in the store
     dispatch(isTxQueued(listItemTx.txStatus) ? addQueuedTransactions(payload) : addHistoryTransactions(payload))
   }, [fetchedTx, chainId, dispatch])
+
+  if (!liveTx && error) {
+    return (
+      <FetchError
+        text="Transaction not found"
+        buttonText="Go to Queue"
+        redirectRoute={SAFE_ROUTES.TRANSACTIONS_QUEUE}
+      />
+    )
+  }
 
   if (!liveTx) {
     return (
